# swagger.json
{
    "info": {
        "description": "API ANTARES",
        "title": "API ANTARES",
        "version": "0.0.1"
    },
    "openapi": "3.0.0",
    "paths": {
        "/file/{path}": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "parameters": [
                {
                    "in": "path",
                    "name": "path",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                },
                {
                    "in": "query",
                    "name": "depth",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "post": {
                "requestBody": {
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "format": "string",
                                "type": "string"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/studies": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "parameters": [
                {
                    "in": "query",
                    "name": "depth",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "post": {
                "requestBody": {
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "format": "string",
                                "type": "string"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/studies/{uuid}": {
            "delete": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "parameters": [
                {
                    "in": "path",
                    "name": "uuid",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                },
                {
                    "in": "query",
                    "name": "depth",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/studies/{uuid}/copy": {
            "parameters": [
                {
                    "in": "path",
                    "name": "uuid",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                },
                {
                    "in": "query",
                    "name": "dest",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                },
                {
                    "in": "query",
                    "name": "depth",
                    "schema": {
                        "type": "integer"
                    }
                }
            ],
            "post": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
...
        "/studies/{uuid}/export": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {}
                        },
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "parameters": [
                {
                    "in": "path",
                    "name": "uuid",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                },
                {
                    "in": "query",
                    "name": "compact",
                    "schema": {
                        "type": "string"
                    }
                },
                {
                    "in": "query",
                    "name": "depth",
                    "schema": {
                        "type": "integer"
                    }
                }
            ]
        },
...
    },
    "servers": [
        {
            "url": "http://0.0.0.0:8080"
        },
        {
            "url": "http://0.0.0.0:5000"
        },
        {
            "url": "http://0.0.0.0:80"
        }
    ],
...
}